Fixed a lot of string formatting errors. Probably fixes some real bugs, especially on windows32

+++ api/api_gui.cpp
@@ -2657,7 +2657,7 @@ float gui_textWidth(const_char* text){
 
 static Gui *get_gui_maybeclosed(int64_t guinum){
   if (guinum < 0 || guinum > g_highest_guinum){
-    handleError("There has never been a Gui #%d", guinum);
+    handleError("There has never been a Gui #%d", (int)guinum);
     return NULL;
   }
 
@@ -2677,13 +2677,13 @@ static Gui *get_gui(int64_t guinum){
 
   if (gui==NULL){
     if (guinum>=0 && guinum<g_highest_guinum)
-      handleError("Gui #%d has been closed and can not be used.", guinum);
+      handleError("Gui #%d has been closed and can not be used.", (int)guinum);
     return NULL;
   }
   
   if (gui->_widget==NULL) {
     R_ASSERT(gui->_created_from_existing_widget); // GUI's that are not created from an existing widget are (i.e. should be) automatically removed when the QWidget is deleted.
-    handleError("Gui #%d (class %s), created from an existing widget, has been closed and can not be used.", gui->get_gui_num(), gui->_class_name.toUtf8().constData());
+    handleError("Gui #%d (class %s), created from an existing widget, has been closed and can not be used.", (int)gui->get_gui_num(), gui->_class_name.toUtf8().constData());
     delete gui;
     return NULL;
   }
@@ -2726,7 +2726,7 @@ int64_t gui_child(int64_t guinum, const_char* childname){
   QWidget *child = gui->_widget->findChild<QWidget*>(childname);
 
   if (child==NULL){
-    handleError("Could not find child \"%s\" in gui #%d.", childname, guinum);
+    handleError("Could not find child \"%s\" in gui #%d.", childname, (int)guinum);
     return -1;
   }
 
@@ -3445,7 +3445,7 @@ void gui_addTableRows(int64_t table_guinum, int pos, int how_many){
   }else if(how_many<0 && pos>=num_rows+how_many){
     if (pos>num_rows+how_many)
       handleError("gui_addTableRows: Illegal 'pos' argument for Gui #%d. pos: %d, how_many: %d, number of rows: %d. Last legal position: %d",
-                  table_guinum, pos, how_many, num_rows, num_rows+how_many);
+                  (int)table_guinum, pos, how_many, num_rows, num_rows+how_many);
     table->setRowCount(num_rows+how_many);
     
   }else if (how_many > 0){
@@ -3624,7 +3624,7 @@ void gui_add(int64_t parentnum, int64_t childnum, int x1_or_stretch, int y1, int
   if(splitter!=NULL || layout==NULL || y1!=-1) {
 
     if (splitter==NULL && layout==NULL && (y1==-1 || x2==-1 || y2==-1)){
-      handleError("Warning: Parent gui #%d does not have a layout", parentnum);
+      handleError("Warning: Parent gui #%d does not have a layout", (int)parentnum);
       x1_or_stretch = 0;
       y1 = 0;
     }
@@ -3692,7 +3692,7 @@ void gui_add(int64_t parentnum, int64_t childnum, int x1_or_stretch, int y1, int
     } else {
 
       if (x1_or_stretch != -1)
-        handleError("Warning: Parent gui #%d does not have a box layout", parentnum);
+        handleError("Warning: Parent gui #%d does not have a box layout", (int)parentnum);
 
       layout->addWidget(child->_widget);
 
@@ -3717,14 +3717,14 @@ void gui_replace(int64_t parentnum, int64_t oldchildnum, int64_t newchildnum){
 
   QLayout *layout = parent->getLayout();
   if(layout==NULL){
-    handleError("Gui #%d does not have a layout", parentnum);
+    handleError("Gui #%d does not have a layout", (int)parentnum);
     return;
   }
   
   QLayoutItem *old_item = layout->replaceWidget(oldchild->_widget, newchild->_widget);
 
   if(old_item==NULL){
-    handleError("Gui #%d not found in #%d", oldchildnum, parentnum);
+    handleError("Gui #%d not found in #%d", (int)oldchildnum, (int)parentnum);
     return;
   }
 
@@ -3800,7 +3800,7 @@ void gui_close(int64_t guinum){
 
   const char *can_not_be_closed_reason = g_guis_can_not_be_closed.value(guinum);
   if (can_not_be_closed_reason != NULL){
-    handleError("Gui #%d can not be closed.\nReason: %s", guinum, can_not_be_closed_reason);
+    handleError("Gui #%d can not be closed.\nReason: %s", (int)guinum, can_not_be_closed_reason);
     return;
   }
   
@@ -3841,7 +3841,7 @@ bool gui_setParent2(int64_t guinum, int64_t parentgui, bool mustBeWindow){
   if (mustBeWindow){
     bool is_window = gui->_widget->isWindow() || gui->_widget->parent()==NULL;
     if(!is_window){
-      handleError("gui_setParent: Gui #%d is not a window. (className: %s)", guinum, gui_className(guinum));
+      handleError("gui_setParent: Gui #%d is not a window. (className: %s)", (int)guinum, gui_className(guinum));
       return false;
     }
   }
@@ -4172,7 +4172,7 @@ void gui_setLayoutSpacing(int64_t guinum, int spacing, int left, int top, int ri
 
   QLayout *layout = gui->getLayout();
   if (layout==NULL){
-    handleError("Gui #%d doesn't have a layout", guinum);
+    handleError("Gui #%d doesn't have a layout", (int)guinum);
     return;
   }
 
@@ -4191,7 +4191,7 @@ void gui_addLayoutSpace(int64_t guinum, int width, int height, bool grow_horizon
 
   QLayout *layout = gui->getLayout();
   if (layout==NULL){
-    handleError("Gui #%d doesn't have a layout", guinum);
+    handleError("Gui #%d doesn't have a layout", (int)guinum);
     return;
   }
 
@@ -4286,11 +4286,11 @@ void gui_setStaticToplevelWidget(int64_t guinum, bool add){
   if (add){
     
     if (is_included){
-      handleError("Gui #%d is already set as static toplevel widget", guinum);
+      handleError("Gui #%d is already set as static toplevel widget", (int)guinum);
       return;
     }
     if (gui->_widget->window() != gui->_widget){
-      handleError("Gui #%d is not a toplevel widget. (class name: %s)", guinum, gui_className(guinum));
+      handleError("Gui #%d is not a toplevel widget. (class name: %s)", (int)guinum, gui_className(guinum));
       return;
     }
     g_static_toplevel_widgets.push_back(gui->_widget);
@@ -4301,7 +4301,7 @@ void gui_setStaticToplevelWidget(int64_t guinum, bool add){
 #endif
 
     if (false==is_included){
-      handleError("Gui #%d has not been added as a static toplevel widget", guinum);
+      handleError("Gui #%d has not been added as a static toplevel widget", (int)guinum);
       return;
     }
     g_static_toplevel_widgets.remove(g_static_toplevel_widgets.indexOf(gui->_widget));
